Update index.html
Here are the key enhancements I'm adding:
1. Search/Filter Functionality: A search bar will be added to the main list screen, allowing you to instantly filter your checklists by organization name as you type.

2. Checklist Sharing (via Import/Export):
You'll be able to generate a unique Share ID for any of your checklists.

Anyone you give this ID to can then import a copy of that checklist into their own account. This is a great way to share templates or completed assessments.

New modals and buttons have been added to facilitate this sharing process seamlessly.